Improve documentation.
